#9322: Split binary to .hpp and .cpp
