+++ ttnn/cpp/ttnn/common.hpp
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include <cstdint>
+
+#pragma once
+
+namespace ttnn {
+
+constexpr uint8_t DefaultQueueId = 0;
+
+}
