+++ ttnn/cpp/ttnn/operations/copy.hpp
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "ttnn/decorators.hpp"
+#include "ttnn/common.hpp"
 #include "ttnn/operations/core/core.hpp"
 #include "ttnn/operations/eltwise/unary/unary.hpp"
 #include "ttnn/operations/eltwise/unary/device/unary_op.hpp"
@@ -66,7 +67,6 @@ struct Typecast {
         const DataType& output_dtype,
         const std::optional<MemoryConfig>& memory_config_arg = std::nullopt,
         const std::optional<Tensor>& optional_output_tensor = std::nullopt) {
-        constexpr uint8_t DefaultQueueId = 0;
         return operator()(DefaultQueueId, input, output_dtype, memory_config_arg, optional_output_tensor);
     }
