+++ ttnn/cpp/ttnn/operations/eltwise/binary/binary_pybind.hpp
@@ -65,8 +65,8 @@ void bind_binary_operation(py::module& module, const binary_operation_t& operati
                const std::optional<const DataType>& dtype,
                const std::optional<ttnn::MemoryConfig>& memory_config,
                const std::optional<ttnn::Tensor>& output_tensor,
-               const std::optional<FusedActivations>& activations,
-               const std::optional<UnaryWithParam>& input_tensor_a_activation,
+               const std::optional<unary::FusedActivations>& activations,
+               const std::optional<unary::UnaryWithParam>& input_tensor_a_activation,
                const uint8_t& queue_id) -> ttnn::Tensor {
                 return self(queue_id, input_tensor_a, scalar, dtype, memory_config, output_tensor, activations, input_tensor_a_activation);
             },
@@ -88,8 +88,8 @@ void bind_binary_operation(py::module& module, const binary_operation_t& operati
                const std::optional<const DataType>& dtype,
                const std::optional<ttnn::MemoryConfig>& memory_config,
                const std::optional<ttnn::Tensor>& output_tensor,
-               const std::optional<FusedActivations>& activations,
-               const std::optional<UnaryWithParam>& input_tensor_a_activation,
+               const std::optional<unary::FusedActivations>& activations,
+               const std::optional<unary::UnaryWithParam>& input_tensor_a_activation,
                const uint8_t& queue_id) -> ttnn::Tensor {
                 return self(queue_id, input_tensor_a, input_tensor_b, dtype, memory_config, output_tensor, activations, input_tensor_a_activation);
             },
