+++ ttnn/cpp/ttnn/operations/eltwise/binary/common/binary_op_types.hpp
@@ -1,0 +1,28 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+namespace ttnn::operations::binary {
+
+enum class BinaryOpType {
+    ADD,
+    SUB,
+    MUL,
+    GT,
+    LT,
+    LTE,
+    GTE,
+    EQ,
+    NE,
+    SQUARED_DIFFERENCE,
+    BIAS_GELU,
+    LOGADDEXP,
+    LOGICAL_AND,
+    LOGICAL_OR,
+    LDEXP,
+    LOGADDEXP2,
+    DIV_FAST
+};
+}
