+++ ttnn/cpp/ttnn/operations/eltwise/binary/device/binary_device_operation.hpp
@@ -8,6 +8,7 @@
 #include <optional>
 #include <variant>
 
+#include "ttnn/common.hpp"
 #include "ttnn/tensor/tensor.hpp"
 #include "third_party/magic_enum/magic_enum.hpp"
 #include "ttnn/tensor/host_buffer/functions.hpp"
@@ -21,48 +22,16 @@
 #include "ttnn/decorators.hpp"
 #include "ttnn/device_operation.hpp"
 #include "ttnn/types.hpp"
+#include "ttnn/operations/eltwise/binary/common/binary_op_types.hpp"
+#include "ttnn/operations/eltwise/binary/common/binary_op_utils.hpp"
 
 namespace ttnn::operations::binary {
-
-constexpr uint8_t DefaultQueueId = 0;
-enum class BinaryOpType {
-    ADD,
-    SUB,
-    MUL,
-    GT,
-    LT,
-    LTE,
-    GTE,
-    EQ,
-    NE,
-    SQUARED_DIFFERENCE,
-    BIAS_GELU,
-    LOGADDEXP,
-    LOGICAL_AND,
-    LOGICAL_OR,
-    LDEXP,
-    LOGADDEXP2,
-    DIV_FAST
-};
-
-using FusedActivations = std::vector<ttnn::operations::unary::UnaryWithParam>;
-
-namespace utils {
-
-std::map<string, string> get_defines(
-    BinaryOpType op_type,
-    const std::optional<DataType> in_dtype = std::nullopt,
-    const std::optional<DataType> out_dtype = std::nullopt,
-    const std::optional<FusedActivations> fused_activations = std::nullopt,
-    const std::optional<unary::UnaryWithParam> input_tensor_a_activation = std::nullopt);
-
-}  // namespace utils
 
 struct BinaryDeviceOperation {
     struct operation_attributes_t {
         BinaryOpType binary_op_type;
         bool in_place;
-        const std::optional<FusedActivations> activations;
+        const std::optional<unary::FusedActivations> activations;
         const std::optional<unary::UnaryWithParam> input_tensor_a_activation;
         const MemoryConfig memory_config;
         const DataType dtype;
