+++ ttnn/cpp/ttnn/operations/eltwise/binary_backward/device/binary_backward_op.hpp
@@ -12,7 +12,6 @@
 
 namespace ttnn::operations::binary_backward {
 
-constexpr uint8_t DefaultQueueId = 0;
 enum class BinaryBackwardOpType {
     ATAN2_BW,
     EMBEDDING_BW,
