+++ ttnn/cpp/ttnn/operations/eltwise/complex_binary_backward/device/complex_binary_backward_op.hpp
@@ -12,7 +12,6 @@
 
 namespace ttnn::operations::complex_binary_backward {
 
-constexpr uint8_t DefaultQueueId = 0;
 enum class ComplexBinaryBackwardOpType {
     COMPLEX_ADD_BW,
     COMPLEX_SUB_BW,
