+++ ttnn/cpp/ttnn/operations/eltwise/complex_unary/device/complex_unary_op.hpp
@@ -12,7 +12,6 @@
 
 namespace ttnn::operations::complex_unary {
 
-constexpr uint8_t DefaultQueueId = 0;
 enum class ComplexUnaryOpType {
     REAL,
     IMAG,
