+++ ttnn/cpp/ttnn/operations/eltwise/ternary/ternary_composite_op.cpp
@@ -4,6 +4,7 @@
 
 
 #include "ternary_composite_op.hpp"
+#include "ttnn/operations/creation.hpp"
 
 namespace ttnn::operations::ternary{
