+++ ttnn/cpp/ttnn/operations/eltwise/unary/device/unary_program_factory_multicore.hpp
@@ -11,6 +11,7 @@
 #include "tt_metal/detail/util.hpp"
 #include "tt_metal/host_api.hpp"
 #include "unary_op.hpp"
+#include "ttnn/operations/eltwise/unary/common/unary_op_utils.hpp"
 
 namespace ttnn::operations::unary::detail {
