+++ ttnn/cpp/ttnn/operations/eltwise/unary/unary.cpp
@@ -4,6 +4,7 @@
 
 #include "unary.hpp"
 
+#include "ttnn/common.hpp"
 #include "device/unary_op.hpp"
 #include "ttnn/run_operation.hpp"
 #include "ttnn/operations/data_movement/downsample/device/downsample_op.hpp"
