+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -11,7 +11,6 @@
 
 namespace ttnn::operations::unary_backward {
 
-constexpr uint8_t DefaultQueueId = 0;
 enum class UnaryBackwardOpType {
     MUL_BW,
     CLAMP_BW,
